Del tekster opp i tre filer - en for hvert språk
