+++ src/language/tekster_nn.ts
@@ -1,0 +1,1017 @@
+// tslint:disable:max-line-length
+export default {
+  'skolepenger.sidetittel': 'Søknad om skolepenger',
+  'skolepenger.overskrift': 'Søknad om stønad til skolepenger',
+
+  'barnetilsyn.sidetittel': 'Søknad om barnetilsyn',
+  'barnetilsyn.hjelpetekst-åpne.hvilke':
+    'Dette sier regelverket om hvilke barn du kan få stønad til barnetilsyn for',
+  'barnetilsyn.hjelpetekst-innhold.hvilke':
+    'Dette sier regelverket om hvilke barn du kan få stønad til barnetilsyn for' +
+    '<br/><br/>' +
+    'Hovedregelen er at du kan få stønad til barnetilsyn frem til barnet ditt har fullført 4. skoleår. \n' +
+    '<br/><br/>' +
+    'På dette alderstrinnet er barn normalt blitt tilstrekkelig selvhjulpne og modne slik at de klarer seg utenfor skoletiden både i hjemmet og i sitt vanlige nærmiljø i den tiden du er fraværende på grunn av arbeid.\n' +
+    '<br/><br/>' +
+    'I noen tilfeller kan du få stønad til barnetilsyn etter at barnet ditt er ferdig med fjerde skoleår:\n' +
+    '<ul>' +
+    '<li>Hvis du har barn som trenger vesentlig mer pass enn jevnaldrende. Du må dokumentere behovet med uttalelse fra lege, spesialist eller annet helsepersonell.</li>' +
+    '<li>Hvis du har en jobb som medfører at du må være borte fra hjemmet i lengre perioder. For at fraværet skal anses som mer langvarig enn vanlig, må det overstige 10 timer per dag. Du må dokumentere arbeidstiden.</li>' +
+    '<li>Hvis du har uregelmessig arbeidstid, for eksempel om kvelden og natten, skiftarbeid, helgearbeid, turnustjeneste, pendlerforhold og arbeid til sjøs. Du må dokumentere arbeidstiden.</li>' +
+    '</ul>',
+
+  'skjema.hei': 'Hei, [0]!',
+  'skjema.knapp.start': 'Start',
+  'skjema.beskrivelse':
+    'Vi kan orientere og veilede deg om utdanning og yrkesvalg. Vi kan også tilby deg for eksempel kurs eller praksisplass, som vil bidra til at du kommer i arbeid.' +
+    '<br/><br/>Når du er arbeidssøker og søker om eller får overgangstønad må du: ' +
+    '<ol><li><a classname="skjema_lenke" href="https://arbeidssokerregistrering.nav.no/start">registrere deg som arbeidssøker</a> hos NAV, hvis du ikke allerede har gjort det</li>' +
+    '<li><a href="https://arbeidsplassen.nav.no/slik-skriver-du-en-god-cv">registrere CV-en din på nav.no</a> og holde den oppdatert</li>' +
+    '<li>søke etter ledige jobber. <a href="https://veiviserarbeidssoker.nav.no/?situasjon=mistet-jobben">Vi har gode tips om jobbsøking.</a></li>' +
+    '<li>sende meldekort hver 14. dag</li></ol>',
+
+  'skjema.tittel.omarbeidssøker': 'Om deg som arbeidssøker',
+  'skjema.alert.registrert':
+    'Du kan registrere deg etter at du har sendt inn dette skjemaet.',
+  'skjema.send': 'Send skjema',
+  'skjema.feilet': 'Noe gikk galt. Skjema er ikke blitt sendt. Prøv igjen.',
+
+  'skjema.forside.disclaimer':
+    'Jeg er klar over at jeg kan miste retten til overgangsstønad dersom jeg ikke har gitt riktige opplysninger. Jeg er også klar over at jeg må betale tilbake dersom jeg får penger jeg ikke har rett til og som skyldes at jeg har latt være å informere eller gitt feil opplysninger.',
+  'skjema.forside.disclaimer.tittel': 'Vi stoler på deg',
+
+  'skjema.oppsummering.disclaimer':
+    'Les gjennom oppsummeringen før du sender inn søknaden. Hvis du trenger å gjøre endringer, kan du gjøre det.',
+  'skjema.oppsummering.omdeg': 'Om deg som arbeidssøker',
+  'skjema.takk': 'Takk',
+  'skjema.alert.mottatt': 'Opplysningene om deg som arbeidssøker er mottatt',
+
+  'banner.tittel.overgangsstønad': 'Søknad om overgangsstønad',
+  'banner.tittel.barnetilsyn': 'Søknad om barnetilsyn',
+  'banner.tittel.skolepenger': 'Søknad om skolepenger',
+  'banner.tittel.arbeidssøker': 'Enslig mor eller far som er arbeidssøker',
+
+  'språk.bokmål': 'Bokmål',
+  'språk.engelsk': 'English',
+  'språk.nynorsk': 'Nynorsk',
+
+  'svar.ja': 'Ja',
+  'svar.nei': 'Nei',
+
+  'knapp.tilbake': 'Tilbake',
+  'knapp.neste': 'Neste',
+  'knapp.avbryt': 'Avbryt',
+  'knapp.start': 'Start søknad',
+  'knapp.sendSøknad': 'Send søknad',
+
+  'feil.ingentekst.sanity': 'Ingen dynamisk tekst funnet i Sanity.',
+  'feil.ingentekst.app': 'Ingen statisk tekst funnet.',
+  'feil.udefinerttekst': 'Udefinert tekst - tom streng (BM)',
+  'feil.alert': 'Noe gikk galt!',
+  'feil.ugyldigTall.beløp': 'Beløpet må være et gyldig tall',
+
+  'datovelger.periode.feilFormat':
+    'Startsdato må være tidligere enn sluttdato.',
+  'datovelger.periode.likeDatoer':
+    'Valgte datoer for perioden kan ikke være like',
+  'datovelger.fødselsdato': 'Fødselsdato',
+  'datovelger.nårSkalDetteSkje': 'Når skal dette skje?',
+
+  'input.kroner': 'kroner',
+
+  'periode.fra': 'Fra',
+  'periode.til': 'Til',
+  'periode.startdato': 'Startdato',
+  'periode.sluttdato': 'Sluttdato',
+  'periode.lesmer-åpne': 'Husker du ikke nøyaktige datoer?',
+  'periode.lesmer-innhold': 'Legg til omtrentelige datoer.',
+
+  'placeholder.beløp': 'F.eks. 5250.30',
+
+  'side.bekreftelse':
+    'Jeg, [0], bekrefter at jeg vil gi riktige og fullstendige opplysninger',
+  'side.info.overgangsstønad':
+    'Overgangsstønaden skal bidra til å sikre inntekt i en tidsbegrenset periode. Vi vil beregne overgangsstønaden din ut i fra den arbeidsinntekten du har eller kan forventes å få.',
+  'side.fortsettSøknad.påbegyntSøknad':
+    'Du har en påbegynt søknad. Vil du fortsette på denne eller starte en ny?',
+  'side.fortsettSøknad.knapp.startPåNytt': 'Start på nytt',
+  'side.fortsettSøknad.knapp.fortsett': 'Fortsett på søknaden',
+
+  'person.navn': 'Navn',
+  'barn.ident': 'Fødselsnummer',
+  'barn.checkbox.ident': 'Barnet har ikke fått fødselsnummer ennå',
+  'person.fødselsdato': 'Fødselsdato (kun hvis du vet)',
+  'person.fnr': 'Fødselsnummer',
+  'person.ident': 'Fødselsnummer (11 siffer) / d-nummer',
+  'person.ident.visning': 'Fødselsnummer eller d-nummer',
+  'person.checkbox.ident': 'Jeg kjenner ikke fødselsnummer / d-nummer',
+  'person.feilmelding.ident': 'Ugyldig fødselsnummer eller d-nummer',
+  'person.telefonnr': 'Telefonnummer du kan kontaktes på',
+  'person.statsborgerskap': 'Statsborgerskap',
+  'person.adresse': 'Adresse',
+  'person.alder': 'Alder',
+
+  barnet: 'Barnet',
+  'ufødt.barn': 'barn med termin [0]',
+  'født.barn': 'barn født [0]',
+  'barnet.storForBokstav': 'Barnet',
+  'barnet.litenForBokstav': 'barnet',
+
+  'stegtittel.omDeg': 'Om deg',
+
+  'personopplysninger.fnr': 'Fødselsnr',
+  'personopplysninger.telefonnr': 'Telefonnummer',
+  'personopplysninger.feilmelding.telefonnr':
+    'Telefonnummeret må ha minst 8 siffer',
+  'personopplysninger.statsborgerskap': 'Statsborgerskap',
+  'personopplysninger.alert.infohentet':
+    'Hvis opplysningene vi har om deg ikke stemmer, må du endre disse hos Folkeregisteret.',
+
+  'personopplysninger.spm.riktigAdresse': 'Bor du på denne adressen?',
+  'personopplysninger.lesmer-åpne.riktigAdresse':
+    'Hvis du har strengt fortrolig adresse',
+  'personopplysninger.lesmer-innhold.riktigAdresse':
+    'Hvis du har strengt fortrolig adresse, vises ikke adressen din her. Du kan svare Ja på dette spørsmålet og fortsette med den digitale søknaden.',
+  'personopplysninger.alert.riktigAdresse':
+    'Du må oppgi riktig adresse til Folkeregisteret for å bruke denne søknaden',
+  'personopplysninger.info.endreAdresse':
+    'Skal du ikke endre adresse i Folkeregisteret?',
+  'personopplysninger.lenke.pdfskjema': 'Bruk PDF-skjema',
+  'personopplysninger.info.pdfskjema':
+    'Skjemaet kan sendes inn elektronisk eller på papir',
+
+  'sivilstatus.tittel': 'Sivilstatus',
+  'sivilstatus.spm.søktSeparasjon':
+    'Har dere søkt om separasjon, søkt om skilsmisse eller reist sak for domstolen?',
+  'sivilstatus.alert-info.søktSeparasjon':
+    'Du må legge bekreftelse fra Fylkesmannen eller domstolen',
+  'sivilstatus.alert-advarsel.søktSeparasjon':
+    'Når du er gift, har du ikke rett til stønad til enslig mor eller far',
+  'sivilstatus.alert.samlivsbruddForeldre':
+    'Du må legge ved erklæring om samlivsbrudd',
+  'sivilstatus.datovelger.søktSeparasjon': 'Når søkte dere eller reiste sak?',
+  'sivilstatus.datovelger.samlivsbrudd': 'Dato for samlivsbrudd',
+  'sivilstatus.datovelger.flyttetFraHverandre':
+    'Når flyttet dere fra hverandre?',
+  'sivilstatus.datovelger.endring':
+    'Når skjedde endringen / når skal endringen skje?',
+  'sivilstatus.spm.erUformeltGift':
+    'Du er registrert som ugift i folkeregisteret. Er du gift uten at det er registrert i folkeregisteret i Norge?',
+  'sivilstatus.alert.erUformeltGift':
+    'Du må legge ved dokumentasjon på inngått ekteskap',
+  'sivilstatus.lesmer-åpne.søkerGiftIUtlandet': 'Hvorfor spør vi om dette?',
+  'sivilstatus.lesmer-innhold.søkerGiftIUtlandet':
+    'Vi henter opplysningene våre fra Folkeregisteret. Hvis du har giftet deg etter religiøse eller kulturelle tradisjoner og ekteskapet ikke er godkjent etter norsk ekteskapslov, er ikke giftemålet registrert i Folkeregisteret.',
+  'sivilstatus.dok.søkerGiftIUtlandet':
+    'Du må legge ved dokumentasjon på inngått ekteskap',
+  'sivilstatus.lesmer-åpne.erUformeltGift': 'Hvorfor spør vi om dette?',
+  'sivilstatus.lesmer-innhold.erUformeltGift':
+    'Vi henter opplysningene våre fra Folkeregisteret. Hvis du har giftet deg etter religiøse eller kulturelle tradisjoner og ekteskapet ikke er godkjent etter norsk ekteskapslov, er ikke giftemålet registrert i Folkeregisteret. ',
+
+  'sivilstatus.spm.erUformeltSeparertEllerSkilt':
+    'Er du separert eller skilt uten at dette er registrert i folkeregisteret i Norge?',
+  'sivilstatus.alert.erUformeltSeparertEllerSkilt':
+    'Du må legge ved dokumentasjon på separasjon eller skilsmisse',
+  'sivilstatus.spm.begrunnelse': 'Hvorfor er du alene med barn?',
+
+  'sivilstatus.svar.samlivsbruddForeldre':
+    'Samlivsbrudd med den andre forelderen',
+  'sivilstatus.svar.samlivsbruddAndre': 'Samlivsbrudd med noen andre',
+  'sivilstatus.svar.aleneFraFødsel': 'Jeg er alene med barn fra fødsel',
+  'sivilstatus.svar.endringISamværsordning': 'Endring i omsorgen for barn',
+  'sivilstatus.svar.dødsfall': 'Jeg er alene med barn på grunn av dødsfall',
+  'sivilstatus.alert.dødsfall':
+    'Når du er alene med barn på grunn av dødsfall, kan du ha rett til stønad til <a href="https://www.nav.no/no/person/pensjon/andre-pensjonsordninger/ytelser-til-gjenlevende-ektefelle" target="_blank" >gjenlevende</a> og <a href="https://www.nav.no/no/person/pensjon/andre-pensjonsordninger/barnepensjon" target="_blank" >barnepensjon</a>. ',
+  'sivilstatus.alert.samlivsbrudd':
+    'Hvis dere ikke er separert eller skilt, må du legge ved bekreftelse på samlivsbruddet. Bekreftelsen må være signert av dere begge og vise dato for bruddet. Du får lenke til et forslag til hvordan bekreftelsen kan se ut når du sender inn søknaden. Dette er ikke meklingsattesten fra familievernkontoret.',
+  'sivilstatus.begrunnelse.endring':
+    'Når skjedde endringen / når skal endringen skje?',
+  'sivilstatus.hjelpetekst-åpne.begrunnelse': 'Grunnen til at vi spør om dette',
+  'sivilstatus.hjelpetekst-innhold.begrunnelse':
+    'Vi spør om dette for å vite hvilken informasjon vi trenger fra deg.',
+
+  'medlemskap.spm.opphold': 'Oppholder du deg i Norge?',
+  'medlemskap.alert-advarsel.opphold':
+    'Når du ikke oppholder deg i Norge, har du som hovedregel ikke rett på stønad til enslig mor eller far. Du kan likevel få stønad dersom du oppholder deg i utlandet fordi du arbeider for en norsk arbeidsgiver. Du kan også være i utlandet i inntil 6 uker i løpet av en 12 måneders periode.',
+  'medlemskap.spm.bosatt': 'Har du bodd i Norge de siste tre årene?',
+  'medlemskap.hjelpetekst-åpne.bosatt': 'Hva vil det si å være bosatt i Norge?',
+  'medlemskap.hjelpetekst-innhold.bosatt':
+    'Hvis du har studert i utlandet med støtte fra lånekassen i denne perioden, regnes du som bosatt i Norge.',
+  'medlemskap.periodeBoddIUtlandet.utenlandsopphold': 'Utenlandsperiode',
+
+  'medlemskap.periodeBoddIUtlandet.slett': 'Fjern utenlandsperiode',
+  'medlemskap.periodeBoddIUtlandet': 'Når bodde du i utlandet?',
+  'medlemskap.periodeBoddIUtlandet.begrunnelse': 'Hvorfor bodde du i utlandet?',
+  'medlemskap.periodeBoddIUtlandet.flereutenlandsopphold':
+    'Har du hatt flere utenlandsopphold de siste 3 årene?',
+  'medlemskap.periodeBoddIUtlandet.knapp': 'Legg til et utenlandsopphold',
+
+  'stegtittel.bosituasjon': 'Bosituasjonen din',
+
+  'bosituasjon.spm.delerSøkerBoligMedAndreVoksne':
+    'Deler du bolig med andre voksne?',
+
+  'bosituasjon.svar.borAleneMedBarnEllerGravid':
+    'Nei, jeg bor alene med barn eller jeg er gravid og bor alene',
+  'bosituasjon.svar.borMidlertidigFraHverandre':
+    'Ja, men jeg og den andre forelderen bor midlertidig fra hverandre',
+  'bosituasjon.svar.borSammenOgVenterBarn':
+    'Ja, jeg bor sammen med en jeg har eller venter barn med',
+  'bosituasjon.svar.harEkteskapsliknendeForhold':
+    'Ja, jeg har samboer og lever i et ekteskapslignende forhold',
+  'bosituasjon.svar.delerBoligMedAndreVoksne':
+    'Ja, jeg deler bolig med andre voksne, for eksempel utleier, venn, søsken eller egne foreldre',
+  'bosituasjon.svar.tidligereSamboerFortsattRegistrertPåAdresse':
+    'Nei, men en tidligere samboer er fortsatt registrert på adressen min',
+
+  'bosituasjon.alert.borMidlertidigFraHverandre':
+    'Når dere bor midlertidig fra hverandre, har du ikke rett til stønad til enslig mor eller far.',
+  'bosituasjon.alert.borSammenOgVenterBarn':
+    'Når du bor sammen med en du har eller venter barn med, har du ikke rett til stønad til enslig mor eller far',
+  'bosituasjon.alert.harEkteskapsliknendeForhold':
+    'Når du har samboer, har du ikke rett til stønad til enslig mor eller far',
+
+  'bosituasjon.tittel.omSamboerenDin': 'omSamboerenDin',
+  'bosituasjon.tittel.omSamboer': 'Om samboeren din',
+  'bosituasjon.tittel.omTidligereSamboer': 'Om den tidligere samboeren din',
+  'bosituasjon.datovelger.nårFlyttetDereSammen': 'Når flyttet dere sammen?',
+  'bosituasjon.datovelger.nårFlyttetDereFraHverandre':
+    'Når flyttet dere fra hverandre?',
+
+  'bosituasjon.alert.tidligereSamboerFortsattRegistrertPåAdresse':
+    '<strong>Vi legger vekt på folkeregistrert adresse. Hvis den tidligere samboeren din ikke samarbeider om adresseendring, kan du kontakte Folkeregisteret for å få hjelp til å endre adressen deres. </strong> <br/>' +
+    '<br/>Du kan også dokumentere at du bor alene med barn og tidspunktet for når dere flyttet fra hverandre ved å legge ved:<br/>' +
+    '<ul>' +
+    '<li>Dokumentasjon som bekrefter at du har vært i kontakt med Folkeregisteret fordi den tidligere samboeren din har flyttet ut</li>' +
+    '<li>Informasjon om hvor den tidligere samboeren din bor nå</li></ul>' +
+    'Hvis du kan dokumentere begge punktene over, er dette som regel tilstrekkelig. Dersom du ikke har slik dokumentasjon, kan du for eksempel legge ved:' +
+    '<ul>' +
+    '<li>Husleiekontrakt for begge parter</li>' +
+    '<li> Dokumentasjon på at tidligere felles bolig er overdratt til en av partene. Dette kan være skifte og/eller dokumentasjon på hvem som er låntaker for boligen du bor i.</li>' +
+    '<li>Bekreftelse fra barnevernet på at du bor alene med barn</li>' +
+    '<li>Dokumentasjon på separate bo- og husholdningsutgifter. Dette kan være kontoutskrifter som viser betalt husleie, eller andre faste boutgifter slik som strøm og kommunale avgifter.</li>' +
+    '</ul>' +
+    'Dersom barnehage, skole, helsestasjon eller lignende kjenner familiesituasjonen din godt, kan en uttalelse fra dem være en del av helhetsvurderingen vi gjør av bosituasjonen din. Dette kan inkludere informasjon om hvem som henter og leverer, og kommer eventuelt i tillegg til annen dokumentasjon.',
+
+  'bosituasjon.spm.skalSøkerGifteSegMedSamboer':
+    'Har du konkrete planer om å gifte deg eller bli samboer?',
+  'bosituasjon.tittel.hvemSkalSøkerGifteEllerBliSamboerMed':
+    'Hvem skal du gifte deg med eller bli samboer med?',
+
+  'barnadine.sidetittel': 'Barna dine',
+  'barnadine.infohentet':
+    'Informasjonen er hentet fra Folkeregisteret. Hvis noe ikke stemmer, kan du endre informasjonen hos Folkeregisteret.',
+
+  'barnadine.hjelpetekst.åpne': 'Hvilke barn kan du få stønad for?',
+  'barnadine.hjelpetekst.innhold': 'Lorem ipsum hjelpetekst jadda',
+  'barnadine.leggtil': 'Legg til barn',
+  'barnadine.label.skalHaBarnepass': 'Skal barnet ha barnepass?',
+  'barnadine.knapp.fjern': 'Fjern fra søknad',
+  'barnadine.knapp.søkBarnetilsyn': 'Søk om stønad til barnetilsyn for barnet',
+  'barnadine.personnummer': 'Personnummer',
+  'barnadine.leggtil.info':
+    'Er du gravid eller har du nylig fått barn som foreløpig ikke er registrert i Folkeregisteret?',
+  'barnadine.leggtil.info.barnetilsyn':
+    'Har du overtatt foreldreansvar for barn uten at dette ennå er registrert i Folkeregisteret?',
+  'barnadine.advarsel.borikke':
+    'Når barnet ikke bor hos deg, har du ikke rett til stønad til enslig mor eller far',
+  'barnadine.advarsel.skalikkebo':
+    'Når barnet ikke skal bo hos deg, har du ikke rett til stønad til enslig mor eller far',
+  'barnadine.spm.borBarnHosDeg': 'Bor barnet hos deg?',
+  'barnadine.info.terminbekreftelse': 'Du må legge ved terminbekreftelse',
+  'barnadine.termindato': 'Termindato',
+  'barnekort.termindato': 'TERMINDATO',
+  'barnekort.fødselsdato': 'FØDSELSDATO',
+  'barnekort.fødselsnummer': 'FØDSELSNUMMER',
+  'barnekort.alder': 'ALDER',
+  'barnekort.bosted': 'BOSTED',
+  'barnekort.adresse.registrert': 'Registrert på adressen din',
+  'barnekort.adresse.uregistrert': 'Ikke registrert på adressen din',
+
+  'barnekort.år': 'år',
+  'barnekort.lenke.endre': 'Endre',
+  'barnekort.normaltekst.barn': 'Barn',
+  'barnekort.spm.skalBarnetBoHosSøker': 'Skal barnet bo hos deg?',
+  'barnekort.spm.født': 'Er barnet født?',
+  'barnekort.erUfødt': 'Ufødt',
+  'barnekort.spm.sammeAdresse': 'Har barnet samme adresse som deg?',
+  'barnekort.født': 'Er barnet født?',
+  'barnekort.adresse.bor': 'Bor hos deg',
+  'barnekort.adresse.borIkke': 'Bor ikke hos deg',
+  'barnekort.adresse.skalBo': 'Skal bo hos deg',
+  'barnekort.adresse.skalIkkeBo': 'Skal ikke bo hos deg',
+  'barnekort.fjern': 'Fjern fra søknad',
+  'barnekort.skalHaBarnepass': 'Skal barnet være med i søknaden?',
+
+  'barnasbosted.kanikkeoppgiforelder':
+    'Jeg kan ikke oppgi den andre forelderen',
+  'barnasbosted.sidetittel': 'Den andre forelderen og samvær',
+  'barnasbosted.borinorge': 'Bor [0]s andre forelder i Norge?',
+  'barnasbosted.hvilketLand': 'Hvilket land bor den andre forelderen i?',
+  'barnasbosted.land': 'Land',
+  'barnasbosted.spm.hvorforikkeoppgi':
+    'Hvorfor kan du ikke oppgi den andre forelderen?',
+  'barnasbosted.spm.donorbarn': 'Donorbarn',
+  'barnasbosted.spm.annet': 'Annet',
+  'barnasbosted.avtale':
+    'Har du og den andre forelderen skriftlig avtale om delt bosted for [0]?',
+  'barnasbosted.alert-advarsel.avtaleDeltBosted':
+    'Når dere har avtale om delt bosted, har ingen av dere rett til stønad som enslig mor eller far for dette barnet.',
+  'barnasbosted.alert-info.avtaleOmDeltBosted':
+    'Du må legge ved avtale om delt bosted',
+  'barnasbosted.alert.måBoHosDeg':
+    '[0] må bo hos deg for at du skal ha rett til stønad',
+  'barnasbosted.spm.skalBarnetBoHosSøker': 'Skal [0] ha adresse hos deg?',
+  'barnasbosted.spm.jaFolkeregistrert':
+    'Ja, og vi har eller skal registrere adressen i Folkeregisteret',
+  'barnasbosted.spm.jaMenSamarbeiderIkke':
+    'Ja, men den andre forelderen samarbeider ikke om adresseendring',
+  'barnasbosted.alert.hvisFaktiskBor':
+    'Hvis barnet faktisk bor hos deg, må du ta kontakt med Folkeregisteret for å få hjelp til å registrere riktig adresse.<br/>' +
+    '\n' +
+    '<br/>Hvis Folkeregisteret ikke kan hjelpe deg, kan du kontakte Familievernkontoret for råd og veiledning.<br/>' +
+    '\n' +
+    '<br/>Du kan også dokumentere at barnet bor hos deg ved å sende inn for eksempel:<br/>' +
+    '\n <ul>' +
+    '<li>grunnen til manglende adresseendring for barnet</li>' +
+    '<li>kopi av flyttemelding/tips til Folkeregisteret</li>' +
+    '<li>bekreftelse fra for eksempel barnehage/skole, barnevern eller helsestasjon</li>' +
+    '</ul>',
+
+  'barnasbosted.alert.skalBarnetBoHosSøker.ja':
+    'Du må oppdatere adressen i Folkeregisteret så fort som mulig, slik at vi kan behandle søknaden din med riktig informasjon.',
+  'barnasbosted.alert.skalBarnetBoHosSøker.nei':
+    'Når barnet ikke bor hos deg, har du ikke rett til stønad til enslig mor eller far.',
+
+  'barnasbosted.element.barnet': 'Barnet',
+  'barnasbosted.element.andreforelder': 's andre forelder',
+  'barnasbosted.forelder.annen': 'Annen forelder',
+  'barnasbosted.forelder.sammesom': 'Samme som',
+  'barnasbosted.knapp.endre': 'Endre informasjon',
+  'barnasbosted.hjelpetekst.bosted.apne':
+    ' Avtale om delt bosted sier ikke kun noe om hvor mye tid barnet bor hos hver av dere',
+  'barnasbosted.hjelpetekst.bosted.innhold':
+    'Foreldre kan velge å inngå en avtale om delt bosted for barnet/barna. Dette er en juridisk avtale i henhold til barneloven §36 som sier noe om hvilke avgjørelser dere må ta sammen - ikke hvor mye tid barnet er hos hver av dere.<br/>' +
+    '\n' +
+    '<br/>En skriftlig avtale om delt bosted innebærer at begge foreldre i fellesskap tar større avgjørelser på vegne av barnet, for eksempel hvor barnet skal gå i barnehage og hvor i landet barnet skal bo.<br/>' +
+    '\n' +
+    '<br/>Når dere har en skriftlig avtale om delt bosted, regnes ingen av dere for å ha aleneomsorg for barnet/barna. Dette gjelder selv om en av dere har klart mer av den daglige omsorgen.',
+  'barnasbosted.spm.harAnnenForelderSamværMedBarn':
+    'Har den andre forelderen samvær med [0]?',
+  'barnasbosted.spm.jaIkkeMerEnnVanlig':
+    'Ja, men ikke mer enn vanlig samværsrett',
+  'barnasbosted.spm.jaMerEnnVanlig': 'Ja, mer enn vanlig samværsrett',
+  'barnasbosted.spm.andreForelderenSamværNei': 'Nei',
+  'barnasbosted.hjelpetekst.samvær.apne': 'Dette er vanlig samværsrett',
+  'barnasbosted.hjelpetekst.samvær.innhold':
+    'Vanlig samværsrett vil si samvær én ettermiddag i uken med overnatting, samvær annenhver helg, til sammen tre uker i sommerferien og annenhver høst-, jule-, vinter- og påskeferie.' +
+    '<br /><br /> Hvis dere har en ordning som tilsvarer denne mengden samvær, men som foregår på andre dager, likestilles det med vanlig samværsrett.' +
+    '<br /><br />Hvis den andre forelderen treffer barnet sjelden og/eller under tilsyn, regnes dette også som ikke mer enn vanlig samværsrett.',
+
+  'barnasbosted.spm.ja': 'Ja',
+  'barnasbosted.spm.nei': 'Nei',
+
+  'barnasbosted.spm.harDereSkriftligSamværsavtale':
+    'Har dere skriftlig samværsavtale for [0]?',
+  'barnasbosted.hjelpetekst-åpne.harDereSkriftligSamværsavtale':
+    'Hva må en skriftlig samværsavtale inneholde?',
+  'barnasbosted.hjelpetekst-innhold.harDereSkriftligSamværsavtale':
+    'Vi trenger opplysninger om' +
+    '<ul><li>hvor mange dager og netter barnet oppholder barnet seg hos den andre forelderen i minst en to ukers-periode</li>' +
+    '<li>når barnet reiser til og fra den andre forelderen</li></ul>',
+  'barnasbosted.spm.jaKonkreteTidspunkt':
+    'Ja, og den beskriver når barnet er sammen med hver av foreldrene',
+  'barnasbosted.spm.jaIkkeKonkreteTidspunkt':
+    'Ja, men den beskriver ikke når barnet er sammen med hver av foreldrene',
+  'barnasbosted.alert.leggeVedSamværsavtalen': 'Du må legge ved samværsavtalen',
+  'barnasbosted.element.samvær': 'Hvordan praktiserer dere samværet?',
+  'barnasbosted.normaltekst.opplysninger': 'Vi trenger opplysninger om',
+  'barnasbosted.normaltekst.hvormangedager':
+    'hvor mange dager og netter barnet oppholder seg hos den andre forelderen i minst en to ukers-periode',
+  'barnasbosted.normaltekst.nårreiserbarnet':
+    'når barnet reiser til og fra den andre forelderen',
+
+  'barnasbosted.spm.borAnnenForelderISammeHus':
+    'Bor du og den andre forelderen til [0] i samme hus, blokk, gårdstun, kvartal eller vei/gate?',
+  'barnasbosted.spm.borAnnenForelderISammeHusBeskrivelse':
+    'Hvordan bor dere nærme hverandre?',
+  'barnasbosted.spm.vetikke': 'Jeg vet ikke hvor den andre forelderen bor',
+  'barnasbosted.hjelpetekst.borAnnenForelderISammeHus.apne':
+    'Grunnen til at vi spør om dette',
+  'barnasbosted.hjelpetekst.borAnnenForelderISammeHus.innhold':
+    'Når du bor svært nær den andre forelderen, regnes du ikke for å være alene om omsorgen for barn.',
+
+  'barnasbosted.spm.boddsammenfør':
+    'Har du bodd sammen med den andre forelderen til [0] før?',
+  'barnasbosted.normaltekst.nårflyttetfra': 'Når flyttet dere fra hverandre?',
+  'barnasbosted.spm.hvorMyeSammen':
+    'Hvor mye er du sammen med den andre forelderen til [0]?',
+  'barnasbosted.lesmer-åpne.hvorMyeSammen': 'Grunnen til at vi spør om dette',
+  'barnasbosted.lesmer-innhold.hvorMyeSammen':
+    'Vi spør om dette for å kunne vurdere om du er så mye sammen med den andre av barnets foreldre at du ikke regnes som enslig mor eller far.',
+  'barnasbosted.spm.møtesIkke': 'Vi møtes ikke',
+  'barnasbosted.spm.kunNårLeveres':
+    'Vi møtes kun når barnet skal hentes eller leveres',
+  'barnasbosted.spm.møtesUtenom': 'Vi møtes også utenom henting og levering',
+  'barnasbosted.spm.beskrivSamværUtenBarn':
+    'Hvor mye er du sammen med den andre forelderen til [0]?',
+  'barnasbosted.skalBliFolkeregistrert.tekst':
+    'Husk å registrere riktig adresse for [0] i Folkeregisteret.',
+  'barnasbosted.skalBliFolkeregistrert.knapp':
+    'Registrer adresse hos Folkeregisteret',
+
+  'stegtittel.arbeidssituasjon': 'Arbeid, utdanning og andre aktiviteter',
+  'stegtittel.arbeidssituasjon.barnetilsyn': 'Arbeidssituasjonen din',
+  'erDuIArbeid.spm': 'Er du i arbeid?',
+  'erDuIArbeid.hjelpetekst':
+    'Hvis du ikke er i arbeid, men er registrert som arbeidssøker hos NAV eller tar utdanning som NAV har godkjent, har du som hovedregel ikke rett til denne stønaden. Da kan du i stedet søke om <a href="https://www.nav.no/familie/alene-med-barn/tilleggsstonader#pass-av-barn">stønad til pass av barn</a>.',
+  'erDuIArbeid.svar.nei': 'Nei, jeg er ikke i arbeid fordi jeg er syk ',
+  'erDuIArbeid.alertsstripe-info':
+    'Du kan få stønaden i inntil ett år når du har en sykdom som ikke er varig, og sykdommen gjør at du ikke kan være i arbeid.<br/><br/>' +
+    'At sykdommen ikke er varig betyr i denne sammenhengen at den ikke har vart i mer enn 2 år eller vil vare i mer enn 2 år. <br/><br/>' +
+    'Det er en forutsetning at barnepassordningen ble opprettet fordi du var i eller skulle begynne i arbeid, og at den fortsetter i den perioden du er syk. ',
+  'erDuIArbeid.alertsstripe-dokumentasjon':
+    '<b>Vi trenger dokumentasjon fra legen din som viser:</b> <ul><li>grunnen til at du ikke kan være i yrkesrettet aktivitet</li><li>når du ble syk</li><li>når legen din regner med at du vil bli frisk</li></ul>' +
+    'Du får muligheten til å laste ned eller skrive ut en huskeliste du kan ta med til legen din for å dokumentere dette når du sender inn søknaden.',
+
+  'arbeidssituasjon.spm': 'Hvordan er arbeidsituasjonen din?',
+  'arbeidssituasjon.spm.hjelpetekst': 'Du kan velge flere alternativer',
+  'arbeidssituasjon.svar.erHjemmeMedBarnUnderEttÅr':
+    'Jeg er hjemme med barn under 1 år',
+  'arbeidssituasjon.svar.erArbeidstaker':
+    'Jeg er arbeidstaker eller lønnsmottaker som frilanser',
+  'arbeidssituasjon.svar.erSelvstendigNæringsdriveneEllerFrilanser':
+    'Jeg er selvstendig næringsdrivende eller frilanser med eget firma',
+  'arbeidssituasjon.svar.erAnsattIEgetAS':
+    'Jeg er ansatt i eget aksjeselskap (AS)',
+  'arbeidssituasjon.svar.etablererEgenVirksomhet':
+    'Jeg etablerer egen virksomhet',
+  'arbeidssituasjon.svar.erArbeidssøker': 'Jeg er arbeidssøker',
+  'arbeidssituasjon.svar.tarUtdanning': 'Jeg tar eller skal ta utdanning',
+  'arbeidssituasjon.svar.erHverkenIArbeidUtdanningEllerArbeidssøker':
+    'Jeg er ikke i arbeid, utdanning eller arbeidssøker',
+  'arbeidssituasjon.svar.harFåttJobbTilbud': 'Jeg har fått tilbud om jobb',
+
+  'arbeidssituasjon.alert.aktivitetspliktFraEttÅr':
+    '<strong>Når barnet ditt er under 1 år kan du motta overgangsstønad uten å jobbe, studere eller søke arbeid.</strong> <br/>' +
+    '\n' +
+    '<br/>For å få overgangsstønad etter at barnet ditt har fylt 1 år, må du jobbe, studere eller søke arbeid.<br/>' +
+    '\n' +
+    '<br/>Du kan fylle aktivitetsplikten på én av disse måtene:<br/>' +
+    '\n <ul>' +
+    '<li>være i minst 50 prosent arbeid</li>' +
+    '<li>være registrert hos NAV som arbeidssøker</li>' +
+    '<li>være i minst 50 prosent utdanning som NAV har vurdert som nødvendig for at du skal kunne komme i jobb og forsørge deg selv</li>' +
+    '<li>etablere egen virksomhet</li>' +
+    '</ul>' +
+    '<br/><br/>Det vil være fornuftig å starte planleggingen av barnepass og arbeid eller utdanning / opplæring så tidlig som mulig.<br/><br/>' +
+    'Du kan likevel få stønad hvis du ikke kan jobbe, studere eller søke arbeid på grunn av egen sykdom. Dette gjelder også hvis du har sykt barn eller mangler barnepass selv om du har søkt. Det samme gjelder dersom barnet ditt trenger særlig tilsyn på grunn av fysiske, psykiske eller store sosiale problemer.',
+
+  'arbeidsforhold.tittel': 'Om arbeidsforholdet ditt',
+  'arbeidsforhold.tittel.arbeidsgiver': 'Arbeidssted',
+  'arbeidsforhold.tittel.egetAS': 'Eget aksjeselskap',
+  'arbeidsforhold.knapp.slettArbeidsgiver': 'Fjern arbeidssted',
+
+  'arbeidsforhold.label.navn': 'Navn på arbeidssted',
+  'arbeidsforhold.label.arbeidsmengde': 'Hvor mye jobber du?',
+  'arbeidsforhold.label.ansettelsesforhold':
+    'Hva slags ansettelsesforhold har du?',
+  'arbeidsforhold.svar.fast': 'Fast stilling',
+  'arbeidsforhold.svar.midlertidig': 'Midlertidig stilling',
+  'arbeidsforhold.svar.lærling': 'Lærling',
+  'arbeidsforhold.alert.lærling': 'Du må legge ved lærlingkontrakten din',
+  'arbeidsforhold.svar.tilkallingsvakt': 'Tilkallingsvikar eller liknende',
+  'arbeidsforhold.label.sluttdato': 'Har du en sluttdato?',
+  'arbeidsforhold.datovelger.sluttdato': 'Når skal du slutte?',
+  'arbeidsforhold.label.flereArbeidsgivere': 'Har du flere arbeidssteder?',
+  'arbeidsforhold.knapp.leggTilArbeidsgiver': 'Legg til et arbeidssted',
+
+  'egetAS.tittel': 'Om aksjeselskapet ditt',
+  'egetAS.label.flere': 'Har du flere egne aksjeselskap?',
+  'egetAS.knapp.fjern': 'Fjern aksjeselskap',
+  'egetAS.knapp.leggtil': 'Legg til et aksjeselskap',
+  'egetAS.label.navn': 'Navn på aksjeselskapet ditt',
+  'egetAS.label.aksjeselskap': 'Aksjeselskap',
+
+  'firmaer.tittel': 'Om firmaet du driver',
+  'firma.tittel': 'Firma',
+  'firma.label.navn': 'Navn på firma',
+  'firma.label.organisasjonnr': 'Organisasjonsnummer',
+  'firma.feilmelding.organisasjonnr': 'Organisasjonsnummeret må ha 9 siffer',
+  'firma.datovelger.etablering': 'Når etablerte du firmaet?',
+  'firma.label.arbeidsmengde': 'Hvor mye jobber du?',
+  'firma.label.arbeidsuke': 'Hvordan ser arbeidsuken din ut?',
+  'firma.knapp.slett': 'Fjern firma',
+  'firmaer.label.flereFirmaer': 'Har du flere firmaer?',
+  'firmaer.knapp.leggTilFirma': 'Legg til firma',
+
+  'arbeidssituasjon.tittel.etablererEgenVirksomhet':
+    'Om virksomheten du etablerer',
+  'arbeidssituasjon.label.etablererEgenVirksomhet': 'Beskriv virksomheten',
+  'arbeidssituasjon.alert.etablererEgenVirksomhet':
+    'Får du ikke dagpenger mens du etablerer egen virksomhet, må du skaffe næringsfaglig vurdering fra kommunen eller fylkeskommunen. Du kan også bruke en annen faglig kompetanse.<br/><br/>' +
+    'Vi trenger denne vurderingen når vi skal ta stilling til om virksomheten kan godkjennes som yrkesrettet aktivitet.',
+  'arbeidssituasjon.label.arbeidsmengde': 'Hvor mye jobber du?',
+
+  'arbeidssøker.tittel': 'Når du er arbeidssøker',
+  'arbeidssøker.label.registrert': 'Er du registrert som arbeidssøker hos NAV?',
+  'arbeidssøker.alert.forÅHaRett':
+    'Du kan registrere deg etter at du har sendt inn dette skjemaet. Da hjelper vi deg videre til registreringen.',
+  'arbeidssøker.søknad.alert.forÅHaRett':
+    'Du kan registrere deg etter at du har sendt inn denne søknaden. Da hjelper vi deg videre til registreringen.',
+  'arbeidssøker.hjelpetekst-åpne.registrert':
+    'Om å være registrert som arbeidssøker hos NAV',
+  'arbeidssøker.hjelpetekst-innhold.registrert':
+    'For å ha rett til overgangsstønad, må du registrere deg som arbeidssøker på nav.no. Hvis du ikke er registrert, kan du gjøre dette etter at du har sendt inn denne søknaden.',
+  'arbeidssøker.label.villig':
+    'Er du villig til å ta imot tilbud om arbeid eller arbeidsmarkedstiltak?',
+  'arbeidssøker.alert.villig':
+    'Hvis det er helsemessige grunner som hindrer deg fra å ta ethvert arbeid, må du dokumentere det med legeattest',
+  'arbeidssøker.label.senestEnUke':
+    'Kan du begynne i arbeid senest én uke etter at du har fått tilbud om jobb?',
+  'arbeidssøker.hjelpetekst-innhold.kanBegynneInnenEnUke':
+    'Du må være tilgjengelig og raskt kunne ta stilling til tilbud om arbeid eller arbeidsmarkedstiltak. Du må derfor kunne skaffe barnepass på kort varsel.',
+  'arbeidssøker.label.barnepass':
+    'Har du eller kan du skaffe barnepass senest innen en uke etter at du har fått tilbud om jobb eller arbeidsmarkedtiltak?',
+  'arbeidssøker.label.ønsketArbeidssted': 'Hvor ønsker du å søke arbeid?',
+  'arbeidssøker.svar.nærme':
+    'Kun i bodistriktet mitt, ikke mer enn 1 times reisevei hver vei',
+  'arbeidssøker.svar.hvorSomHelst': 'Hvor som helst i landet',
+  'arbeidssøker.lesmer-åpne.ønsketArbeidssted': 'Dette betyr 1 times reisevei',
+  'arbeidssøker.lesmer-innhold.ønsketArbeidssted':
+    '1 times reisevei vil si at reisetiden med transportmiddel mellom hjem og arbeidssted ikke er over 1 time hver vei. Det inkuderer ikke gangtid mellom bolig og transportmiddel, og transportmiddel og arbeidssted.',
+  'arbeidssøker.label.halvstilling':
+    'Ønsker du å stå som arbeidssøker til minst 50 prosent stilling?',
+  'arbeidssøker.alert.halvstilling':
+    'Du kan søke heltidsjobb, men som enslig mor eller far holder det at du jobber minst 50 prosent. Det er heller ikke krav om at du må jobbe kvelds-, natt-, helg- og skiftarbeid.',
+  'arbeidssøker.tekst.tillegstønad':
+    'Som enslig mor eller far som søker arbeid, kan du ha rett til tilleggsstønader <br/> ' +
+    'Stønadene kan dekke utgifter til <ul>' +
+    '<li>barnepass</li>' +
+    '<li>flytting</li>' +
+    '</ul>' +
+    'NB! Du kan kun få tilleggsstønader hvis du kvalifiserer til overgangsstønad',
+  'arbeidssøker.lenke.tilleggstønad': 'Les mer om tilleggstønader',
+  'arbeidssøker.knapp.tilleggstønad': 'Søk tilleggsstønader',
+
+  'stegtittel.utdanning': 'Utdanningen din',
+  'tidligereUtdanning.lesmer-åpne.kanFåStønad':
+    'Grunnen til at vi spør om dette',
+  'tidligereUtdanning.lesmer-innhold.kanFåStønad':
+    'Vi spør om tidligere utdanning for å kunne vurdere om utdanningen du skal ta er nødvendig for at du skal kunne komme i jobb og forsørge deg selv.<br/><br/>' +
+    'Har du allerede en utdanning som gjør deg kvalifisert til et yrke eller som kan benyttes i flere yrker, vil du som hovedregel ikke få støtte til mer utdanning.<br/><br/>' +
+    'Vi kan gjøre unntak hvis du har en utdanning eller yrkeserfaring som ikke lenger er relevant i dagens arbeidsmarked. Høgskole eller universitetsutdanning blir ikke utdatert.<br/><br/>' +
+    'Vi kan også gjøre unntak fra denne regelen hvis utdanningen din/yrket ditt ikke er forenlig med omsorgen for barn. Dette gjelder ikke vanlig turnusarbeid.',
+
+  'utdanning.tittel': 'Utdanningen du tar eller skal ta',
+  'utdanning.undertittel': 'Utdanning',
+  'utdanning.lesmer-åpne.kanFåStønad': 'Utdanning du kan få stønad til',
+  'utdanning.lesmer-innhold.kanFåStønad':
+    'For å få rett til overgangsstønad under utdanning, må NAV vurdere utdanningen din som nødvendig for at du skal kunne komme i jobb og forsørge deg selv.<br/><br/>' +
+    'Da vurderer vi:' +
+    '<ul><li>hva som er målet med utdanningen din</li>' +
+    '<li>om det er nødvendig med utdanning for å kunne nå målet ditt</li>' +
+    '<li>om utdanningen er hensiktsmessig. Da vurderer vi blant annet både arbeidsmarkedets behov og dine muligheter.</li></ul>' +
+    'Når du søker om overgangsstønad vil vi vurdere utdanningen din.',
+  'utdanning.lesmer-innhold.kanFåStønad.skolepenger':
+    'For å få rett til stønad til skolepenger, må denne utdanningen være nødvendig for at du skal kunne komme i jobb og forsørge deg selv. <br /><br />' +
+    'Vi vurderer:' +
+    '<ul><li>hvilket mål du har for å skaffe eller beholde inntektsgivende arbeid</li>' +
+    '<li>om det er nødvendig med utdanning for å kunne nå målet ditt</li>' +
+    '<li>om den aktuelle utdanningen er hensiktsmessig. Da vurderer vi blant annet arbeidsmarkedets behov og dine muligheter.</li>' +
+    '</ul>' +
+    'Vi vurderer utdanningen du vil ta når du søker om stønad. ',
+  'utdanning.label.skoleUtdanningssted': 'Skole / utdanningssted',
+  'utdanning.label.linjeKursGrad': 'Linje / kurs / grad',
+  'utdanning.spm.privatEllerOffentlig':
+    'Er utdanningen offentlig eller privat?',
+  'utdanning.lesmer-åpne.privatEllerOffentlig':
+    'Om godkjenning av privat utdanning',
+  'utdanning.lesmer-innhold.privatEllerOffentlig':
+    'Vi godkjenner kun privat utdanning dersom det offentlige skoleverket ikke har det samme tilbudet.<br/><br/>Vi kan godta privat utdanning dersom det er særlige grunner, for eksempel at du ikke har kommet inn ved offentlig skole eller at du har begynt på den private utdanningen før du ble enslig mor eller far.',
+  'utdanning.svar.offentlig': 'Offentlig',
+  'utdanning.svar.privat': 'Privat',
+  'utdanning.datovelger.studieperiode': 'Når var du elev / student?',
+  'utdanning.datovelger.studieperiode.fremtidig':
+    'Når skal du være elev/student?',
+  'utdanning.spm.studieandel': 'Er utdanningen på heltid eller deltid?',
+  'utdanning.svar.heltid': 'Heltid',
+  'utdanning.svar.deltid': 'Deltid',
+  'utdanning.label.arbeidsmengde': 'Hvor mye skal du studere?',
+  'utdanning.spm.mål': 'Hva er målet med utdanningen?',
+
+  'utdanning.label.utgifter': 'Utgifter til skolepenger',
+  'utdanning.label.utgifter.dokumentasjon':
+    'Dokumentasjon på utgiftene dine til studieavgift, semesteravgift og/eller eksamensgebyr. <br />Det dokumenterer du med faktura i ditt navn.',
+  'utdanning.label.utgifter.eksamensgebyr': 'Eksamensgebyr',
+  'utdanning.label.utgifter.studieavgift': 'Studieavgift',
+  'utdanning.label.utgifter.semesteravgift': 'Semesteravgift',
+
+  'utdanning.tittel.tidligere': 'Tidligere utdanning',
+  'utdanning.spm.grunnskole': 'Har du tatt utdanning etter grunnskolen?',
+  'utdanning.lesmer-åpne.grunnskolen': 'Dette regnes som grunnskolen',
+  'utdanning.lesmer-innholde.grunnskolen':
+    'Grunnskolen er barne- og ungdomsskole, det vil si til og med 10. klasse.',
+  'utdanning.label.leggtil': 'Har du tatt mer utdanning?',
+  'utdanning.knapp.leggtil': 'Legg til en utdanning',
+  'utdanning.knapp.slett': 'Fjern utdanning',
+
+  'utdanning.andreUtgifter.tittel':
+    'Har du andre utgifter i forbindelse med utdanningen?',
+  'utdanning.andreUtgifter.tekst':
+    '<strong>Har du andre utgifter i forbindelse med utdanningen?</strong><br /><br />' +
+    'Tilleggsstønader kan dekke utgifter til:<ul>' +
+    '<li>barnepass</li>' +
+    '<li>læremidler</li>' +
+    '<li>daglig reise</li>' +
+    '<li>reise til samling</li>' +
+    '<li>reise på grunn av oppstart, avslutning eller hjemreise</li>' +
+    '<li>bolig og overnatting</li>' +
+    '<li>flytting</li>' +
+    '</ul>' +
+    'Vi viser deg videre til søknad om tilleggsstønader når du er ferdig med denne søknaden.\n' +
+    'Du kan kun få tilleggsstønader hvis du kvalifiserer til overgangsstønad.',
+  'utdanning.andreUtgifter.fakturatekst':
+    'Du må legge ved faktura i ditt navn som viser utgiftene dine til studieavgift, semesteravgift og/eller eksamensgebyr',
+
+  'stegtittel.dinSituasjon': 'Mer om situasjonen din',
+  'dinSituasjon.spm': 'Gjelder noe av dette deg?',
+  'dinSituasjon.svar.erSyk': 'Jeg er syk',
+  'dinSituasjon.svar.harSyktBarn': 'Barnet mitt er sykt',
+  'dinSituasjon.svar.harSøktBarnepassOgVenterEnnå':
+    'Jeg har søkt om barnepass, men ikke fått plass enda',
+  'dinSituasjon.svar.harBarnMedSærligeBehov':
+    'Jeg har barn som trenger særlig tilsyn på grunn av fysiske, psykiske eller store sosiale problemer',
+  'dinSituasjon.svar.nei': 'Nei',
+
+  'dinSituasjon.datovelger.utdanning': 'Når skal du starte utdanningen?',
+  'dinSituasjon.datovelger.jobb': 'Når skal du starte i ny jobb?',
+
+  'dinSituasjon.spm.sagtOppEllerRedusertStilling':
+    'Har du sagt opp jobben eller redusert arbeidstiden de siste 6 månedene?',
+  'dinSituasjon.lesmer-åpne': 'Grunnen til at vi spør om dette',
+  'dinSituasjon.lesmer-innhold':
+    'Har du sagt opp jobben uten rimelig grunn, kan du som hovedregel først ha rett til stønaden 6 måneder etter at du sa opp. Det samme gjelder hvis du har redusert arbeidstiden til under 50 prosent.',
+
+  'dinSituasjon.svar.sagtOpp':
+    'Ja, jeg har sagt opp jobben eller tatt frivillig permisjon (ikke foreldrepermisjon)',
+  'dinSituasjon.svar.redusertStilling': 'Ja, jeg har redusert arbeidstiden',
+
+  'dinSituasjon.alert.sagtOpp':
+    'Du må legge ved dokumentasjon om arbeidsforholdet og grunnen til at du sluttet. Dokumentasjonen må vise:<br/><br/>' +
+    '<ul><li>arbeidsforholdet</li>' +
+    '<li>grunnen til at du sluttet</li>' +
+    '<li>datoen du sa opp eller avtalte frivillig permisjon</li></ul>',
+  'dinSituasjon.fritekst.sagtOpp': 'Hvorfor sa du opp?',
+  'sagtOppEller.datovelger.sagtOpp': 'Når sa du opp?',
+  'dinSituasjon.datovelger-alert.sagtOpp':
+    'Har du sagt opp jobben uten rimelig grunn, kan du først ha rett til stønaden 6 måneder etter at du sa opp.',
+
+  'dinSituasjon.alert.redusertStilling':
+    'Du må legge ved dokumentasjon på arbeidsforholdet og grunnen til at du reduserte arbeidstiden.<br/><br/>' +
+    'Dokumentasjonen må vise:' +
+    '<ul><li>arbeidsforholdet</li>' +
+    '<li>grunnen til at du reduserte arbeidstiden</li>' +
+    '<li>datoen du avtalte reduksjon i arbeidstiden</li></ul>',
+  'dinSituasjon.fritekst.redusertStilling':
+    'Hvorfor reduserte du arbeidstiden?',
+  'dinSituasjon.datovelger.redusertStilling': 'Når reduserte du arbeidstiden?',
+  'dinSituasjon.datovelger-alert.redusertStilling':
+    'Har du redusert arbeidstiden uten rimelig grunn, kan du først ha rett til stønaden 6 måneder etter at du sa opp.',
+  'dinSituasjon.legend.særligTilsyn':
+    'Hvilket av barna dine har behov for særlig tilsyn?',
+  'dinSituasjon.label.særligTilsyn': 'Om {barnetsNavn} tilsynsbehov',
+  'dinSituasjon.alert-dok.tittel':
+    'Du må legge ved dokumentasjon som bekrefter at barnet ditt har behov for særlig tilsyn.',
+
+  'harBarnMedSærligeBehov.alert-dok.beskrivelse':
+    'Dokumentasjonen fra lege må bekrefte: <ul><li>at barnet har medisinske, psykiske eller store sosiale problemer og trenger tilsyn</li>' +
+    '<li>at dette påvirker muligheten din til å være i yrkesrettet aktivitet</li></ul>',
+
+  'sagtOppEllerRedusertStilling.datovelger.sagtOpp': 'Når sa du opp?',
+  'sagtOppEllerRedusertStilling.datovelger.redusertStilling':
+    'Når reduserte du arbeidstiden?',
+  'sagtOppEllerRedusertStilling.datovelger-alert.sagtOpp':
+    'Har du sagt opp jobben uten rimelig grunn, kan du først ha rett til stønaden 6 måneder etter at du sa opp.',
+  'sagtOppEllerRedusertStilling.datovelger-alert.redusertStilling':
+    'Du må legge ved dokumentasjon på arbeidsforholdet, grunnen til at du reduserte arbeidstiden og datoen du avtalte reduksjon i arbeidstiden',
+
+  'søkerFraBestemtMåned.datovelger': 'Når søker du stønad fra?',
+  'søkerFraBestemtMåned.spm.overgangsstønad':
+    'Søker du overgangsstønad fra en bestemt måned?',
+  'søkerFraBestemtMåned.spm.barnepass':
+    'Søker du barnepass fra en bestemt måned?',
+  'søkerFraBestemtMåned.hjelpetekst-åpne': 'Om å søke fra et bestemt tidspunkt',
+  'søkerFraBestemtMåned.hjelpetekst-innhold.overgangsstønad-del1':
+    'Du kan få overgangsstønad fra og med måneden etter at du har rett til stønaden. ' +
+    'Du kan ha rett til stønad i inntil 3 måneder før du søker. ' +
+    'Det vil si fra og med [0]. ',
+  'søkerFraBestemtMåned.hjelpetekst-innhold.overgangsstønad-del2':
+    'Hvis du er gravid, kan du ha rett til overgangsstønad fra måneden før fødsel. ' +
+    'Hvis du har fått barn i løpet av de siste 3 månedene, kan du få stønad i inntil 5 måneder før du søker. ' +
+    'Det vil si fra og med [0]. ',
+  'søkerFraBestemtMåned.hjelpetekst-innhold.overgangsstønad-del3':
+    'Selv om du søker fra en bestemt måned vil vi vurdere om du har rett til stønad fra denne måneden eller senere.',
+  'søkerFraBestemtMåned.hjelpetekst-innhold.barnepass':
+    'Stønad til barnetilsyn utbetales fra og med den måneden du fyller vilkårene for stønad. Du kan få etterbetalt stønad for inntil 3 måneder fra du søker dersom du har rett til stønad tilbake i tid. Vi vil vurdere fra hvilket tidspunkt du har rett til stønad selv om du søker fra en bestemt måned.',
+  'søkerFraBestemtMåned.svar.neiNavKanVurdere':
+    'Nei, Nav kan vurdere fra hvilken måned jeg har rett til stønad',
+
+  'søkerFraBestemtMåned.datovelger.overgangsstønad':
+    'Jeg søker overgangsstønad fra og med',
+  'søkerStønadFraBestemtMnd.datovelger.barnepass':
+    'Jeg søker stønad til barnetilsyn fra og med',
+  'barnepass.dokumentasjon.søkerStønadFraBestemtMnd':
+    'Husk å legge ved tidligere fakturaer hvis du søker stønad tilbake i tid',
+
+  'barnepass.sidetittel': 'Om barnepassordningen',
+  'barnepass.spm.årsak': 'Hvorfor trenger [0] pass?',
+  'barnepass.alert-advarsel.årsak':
+    '<b>Hovedregelen er at du kan få stønad til barnetilsyn frem til barnet ditt har fullført 4. skoleår. </b> <br/><br/>' +
+    'På dette alderstrinnet er barn normalt blitt tilstrekkelig selvhjulpne og modne slik at de klarer seg utenfor skoletiden både i hjemmet og i sitt vanlige nærmiljø i den tiden du er fraværende på grunn av arbeid.',
+  'barnepass.svar.trengerMerPassEnnJevnaldrede':
+    'Barnet har behov for vesentlig mer pass enn det som er vanlig for jevnaldrende',
+  'barnepass.dokumentasjon.trengerMerPassEnnJevnaldrede':
+    'Du må dokumentere behovet med uttalelse fra lege, spesialist eller annet helsepersonell.',
+
+  'barnepass.svar.myeBortePgaJobb':
+    'Jeg må være borte fra hjemmet i lengre perioder på grunn av jobb',
+  'barnepass.alert-info.myeBortePgaJobb':
+    'For at fraværet skal anses som mer langvarig enn vanlig, må det overstige 10 timer per dag',
+  'barnepass.svar.utenomVanligArbeidstid':
+    'Jeg jobber turnus eller skift, og jobber på tider utenom vanlig arbeidstid',
+  'barnepass.dokumentasjon.arbeidstid': 'Du må dokumentere arbeidstiden din',
+
+  'barnepass.spm.hvaSlagsOrdning': 'Hva slags barnepassordning har [0]?',
+  'hvaSlagsOrdning.svar.barnehageOgLiknende': 'Barnehage, SFO eller liknende',
+  'hvaSlagsOrdning.svar.privat': 'Dagmamma eller annen privat ordning',
+  'barnehageOgLiknende.label.navnPåBarnepass': 'Navn på barnepassordningen',
+  'privat.label.navnPåBarnepass':
+    'Navn på barnepassordningen eller personen som passer [0]',
+  'barnepass.datovelger.periodePåBarnepass':
+    'I hvilken periode har [0] denne barnepassordningen?',
+  'barnepass.hjelpetekst.periodePåBarnepass':
+    'Stønad til barnetilsyn gis normalt for 1 år av gangen, og du må søke på nytt og dokumentere utgiftene hvert år. Du trenger bare oppgi perioden du søker for nå.',
+
+  'barnepass.label.beløp': 'Beløp pr måned (ikke inkludert kost)',
+  'barnepass.hjelpetekst-åpne.beløp': 'Dette kan du få støtte til',
+  'barnepass.hjelpetekst-innhold.beløp':
+    ' Stønaden bidrar til utgifter du har til opphold. Utgifter til kost, bleier og lignende dekkes ikke.',
+  'barnepass.alert-dokumentasjon.beløp.barnehageOgLiknende':
+    '<strong>Du må legge ved faktura fra barnepassordningen.</strong> <br />' +
+    'Vi må ha ny faktura for hvert barnehage-/skoleår. Det er ikke tilstrekkelig at du har levert faktura for tidligere år. Fakturaen må stå i ditt navn, og utgifter til opphold må være spesifisert.',
+  'barnepass.alert-dokumentasjon.beløp.privat':
+    '<b>Du må legge ved avtalen du har med barnepasseren. Her er et <a href="/familie/alene-med-barn/soknad/filer/Avtale_privat_barnepass.pdf" download>forslag til hvordan en privat avtale om barnepass kan se ut</a>. Avtalen må inneholde:</b> <ul>' +
+    '<li>ditt navn, fødseslnummer og adresse</li>' +
+    '<li> barnepasserens navn og adresse</li>' +
+    '<li>navn og fødselsnummer for barnet/barna som passes</li>' +
+    '<li>perioden avtalen gjelder for</li>' +
+    '<li>dager i uken og klokkeslett for barnepass</li>' +
+    '<li>hvor mye du betaler for barnepass per måned, ikke inkludert kostpenger</li>' +
+    '<li>signaturer fra deg og barnepasser </li></ul>',
+  'barnepass.label.leggTilOrdning': 'Har du flere barnepassordninger for [0]?',
+  'barnepass.knapp.leggTilOrdning': 'Legg til en barnepassordning',
+  'barnepass.tittel.ordning': 'Barnepassordning',
+  'barnepass.knapp.slett': 'Fjern barnepassordning',
+
+  'filopplaster.dra': 'Last opp dokumentasjon',
+  'filopplaster.slipp': 'Slipp filen her...',
+  'filopplaster.feilmelding.filtype': '[0] – Ugyldig filtype.',
+  'filopplaster.feilmelding.maks':
+    '[0] er for stor (maksimal filstørrelse er [1]).',
+  'filopplaster.feilmelding.generisk': 'Kunne ikke laste opp filen.',
+
+  'oppsummering.sidetittel': 'Oppsummering',
+  'oppsummering.normaltekst.lesgjennom':
+    'Les gjennom oppsummeringen før du sender inn søknaden. Hvis du trenger å gjøre endringer, kan du gå tilbake.',
+  'oppsummering.tilbake': 'Til oppsummering',
+
+  'dokumentasjon.tittel': 'Last opp dokumentasjon',
+  'dokumentasjon.ingenDokumentasjonsbehov.beskrivelse':
+    'Det ser ikke ut til at du må sende inn noe dokumentasjon.  Hvis vi likevel trenger dokumentasjon fra deg, tar vi kontakt.',
+  'dokumentasjon.beskrivelse':
+    'Du må laste opp dokumentasjon som bekrefter opplysningene i søknaden. ' +
+    'Du får raskere svar på søknaden din hvis vi har all dokumentasjonen når vi starter behandlingen. ' +
+    'Du kan bruke filformatene PDF, jpg og png.' +
+    ' <br/><br/>' +
+    'Slik bruker du bilder som vedlegg til søknaden:<br/>' +
+    '<ol><li> Ta bilde av dokumentet med smarttelefon eller nettbrett</li>' +
+    '<li>Sjekk at dokumentet er lett å lese\n</li>' +
+    '<li>Last opp bildene her</li></ol> ' +
+    'Mangler du noe av dokumentasjonen? Da kan du sende inn det du har og ettersende resten.\n <br/>' +
+    '<br />' +
+    'Har du sendt inn denne dokumentasjonen til NAV tidligere? Da trenger du ikke å sende den på nytt.',
+  'dokumentasjon.checkbox.sendtTidligere':
+    'Jeg har sendt inn denne dokumentasjonen til NAV tidligere',
+
+  'dokumentasjon.inngåttEkteskap.tittel': 'Dokumentasjon på inngått ekteskap',
+  'dokumentasjon.separasjonEllerSkilsmisse.tittel':
+    'Dokumentasjon på separasjon eller skilsmisse',
+  'dokumentasjon.separasjonEllerSkilsmisse.beskrivelse':
+    'Beskrivelse for separasjone eller skilsmisse',
+  'dokumentasjon.søktSeparasjon.tittel':
+    'Dokumentasjon på separasjon eller skilsmisse',
+  'dokumentasjon.søktSeparasjon.beskrivelse':
+    'Bekreftelse fra Fylkesmannen eller domstolen på søknad om separasjon, søknad om skilsmisse eller at det er reist sak for domstolen',
+  'dokumentasjon.begrunnelse.tittel':
+    'Bekreftelse på samlivsbrudd med den andre forelderen',
+  'dokumentasjon.begrunnelse.beskrivelse':
+    'Hvis dere ikke er separert eller skilt, må du legge ved bekreftelse på samlivsbruddet. Bekreftelsen må være signert av dere begge og vise dato for bruddet.' +
+    '<br /><br /> Du får lenke til et forslag til hvordan bekreftelsen kan se ut når du sender inn søknaden slik at du kan ettersende dette hvis du ikke allerede har en bekreftelse.' +
+    '<br /><br /> Dette er ikke meklingsattesten fra familievernkontoret.',
+
+  'dokumentasjon.ikke.villig.til.arbeid.tittel':
+    'Dokumentasjon som beskriver grunnen til at du ikke kan ta ethvert arbeid',
+  'dokumentasjon.ikke.villig.til.arbeid.beskrivelse':
+    'Hvis det er helsemessige grunner som hindrer deg fra å ta ethvert arbeid, må du dokumentere det med legeattest',
+  'dokumentasjon.ulikeAdresser.tittel':
+    'Dokumentasjon på at du og tidligere samboer bor på ulike adresser',
+  'dokumentasjon.ulikeAdresser.beskrivelse':
+    'Dokumentasjon på at du og tidligere samboer bor på ulike adresser, for eksempel: <ul>' +
+    '<li>Informasjon om hvor den tidligere samboeren bor nå</li>' +
+    '<li>Kopi av flyttemelding/tips til folkeregisteret</li>' +
+    '<li>Husleiekontrakt for begge parter</li>' +
+    '<li>Dokumentasjon på at tidligere felles bolig er overdratt til en av partene. Dette kan være skifte og/eller dokumentasjon på hvem som er låntaker for boligen du bor i.</li>' +
+    '<li>Bekreftelse fra eksempel barnehage, skole, barnevern, helsestasjon eller lignende.</li>' +
+    '<li>Dokumentasjon på separate bo- og husholdningsutgifter. Dette kan være kontoutskrifter som viser betalt husleie, eller andre faste boutgifter slik som strøm og kommunale avgifter.</li></ul>',
+
+  'dokumentasjon.terminbekreftelse.tittel': 'Terminbekreftelse',
+  'dokumentasjon.terminbekreftelse.beskrivelse': '',
+
+  'dokumentasjon.samværsavtale.tittel': 'Samværsavtale ',
+  'dokumentasjon.deltBosted.tittel': 'Avtale om delt bosted',
+  'dokumentasjon.barnBorHosSøker.tittel':
+    'Dokumentasjon på at barn bor hos deg',
+  'dokumentasjon.barnBorHosSøker.beskrivelse':
+    ' Dokumentasjon på at barn bor hos deg, for eksempel:<ul>' +
+    '<li>Redegjørelse for årsaken til manglende adresseendring for barnet</li>' +
+    '<li>Kopi av flyttemelding/tips til Folkeregisteret</li>' +
+    '<li>Bekreftelse fra for eksempel barnehage/skole, barnevern eller helsestasjon</li></ul>',
+  'dokumentasjon.sykdom.tittel': 'Dokumentasjon på at du er syk',
+  'dokumentasjon.sykdom.beskrivelse':
+    'Dokumentasjonen fra legen din må tydelig vise:<ul>' +
+    '<li>årsaken til at du ikke kan være i yrkesrettet aktivitet</li>' +
+    '<li>hvor mye du kan arbeide</li> </ul>' +
+    'Søker du om å forlenge stønadsperioden fordi du har en sykdom som ikke er varig? Da trenger vi dokumentasjonen fra legen din selv om du har sykemelding. Denne må vise:\n' +
+    '<ul><li>når du ble syk</li>' +
+    '<li>når legen din regner med at du vil bli frisk</li></ul>' +
+    'Har du sendt inn dokumentasjon på dette tidligere, for eksempel i forbindelse med sykemelding, AAP eller uføretrygd?',
+  'dokumentasjon.syktBarn.tittel': 'Dokumentasjon på barnets sykdom',
+  'dokumentasjon.syktBarn.beskrivelse':
+    'Dokumentasjonen fra lege må tydelig vise:\n' +
+    '<ul><li>årsaken til at barnets sykdom påvirker muligheten din til å være i arbeid eller annen yrkesrettet aktivitet</li>' +
+    '<li>hvor mye kan du arbeide</li></ul>' +
+    'Søker du om å forlenge stønadsperioden fordi barnet har en sykdom som ikke er varig? Da trenger vi dokumentasjonen fra lege som også viser:\n' +
+    '<ul><li>når barnet ble sykt</li>' +
+    '<li>når legen regner med at barnet vil bli friskt</li></ul>',
+  'dokumentasjon.barnepass.tittel': 'Dokumentasjon på at du mangler barnepass ',
+  'dokumentasjon.barnepass.beskrivelse':
+    'Det vil si avslag på barnehageplass/SFO-plass eller bekreftelse på at barnet står på venteliste.' +
+    'Dokumentasjonen må tydelig vise <ul>' +
+    '<li>datoen du søkte</li>' +
+    '<li>datoen du ønsket plass fra</li></ul>',
+  'dokumentasjon.barnetilsynsbehov.tittel':
+    'Dokumentasjon på barnets tilsynsbehov',
+  'dokumentasjon.barnetilsynsbehov.beskrivelse':
+    '<ul><li>Dokumentasjon fra legen din som bekrefter at barnet har medisinske, psykiske eller store sosiale problemer og trenger tilsyn.<li/>' +
+    '<li>Dokumentasjon som beskriver behovet barnet ditt har for tilsyn og hvordan dette påvirker muligheten din til å være i arbeid eller yrkesrettet aktivitet.</li></ul>',
+  'dokumentasjon.arbeidskontrakt.tittel':
+    'Arbeidskontrakt som viser at du har fått tilbud om arbeid.',
+  'dokumentasjon.arbeidskontrakt.beskrivelse':
+    'Dokumentasjonen må tydelig vise: <ul>' +
+    '<li>navn på arbeidssted</li>' +
+    '<li>stillingsprosent</li>' +
+    '<li>dato for oppstart</li>' +
+    '<li>dato du fikk tilbudet</li></ul>',
+  'dokumentasjon.utdanning.tittel':
+    'Dokumentasjon på utdanningen du tar eller skal ta',
+  'dokumentasjon.utdanning.beskrivelse':
+    'Dokumentasjonen må vise:<ul>' +
+    '<li>navn på studiested</li>' +
+    '<li>navn på studie</li>' +
+    '<li>hvor mye du skal studere</li>' +
+    '<li>perioden du skal studere</li></ul>',
+
+  'dokumentasjon.lærling.tittel': 'Lærlingkontrakt',
+
+  'dokumentasjon.arbeidsforhold-oppsigelse.tittel':
+    'Dokumentasjon på arbeidsforholdet og årsaken til at du sluttet',
+  'dokumentasjon.arbeidsforhold-oppsigelse.beskrivelse': '',
+  'dokumentasjon.arbeidsforhold-redusert.tittel':
+    'Dokumentasjon på arbeidsforholdet og årsaken til at du reduserte arbeidstiden',
+  'dokumentasjon.arbeidsforhold-redusert.beskrivelse': '',
+
+  'dokumentasjon.syk.tittel': 'Dokumentasjon på at du er syk',
+  'dokumentasjon.syk.beskrivelse':
+    'Dokumentasjonen fra legen din må tydelig vise: <b/><b/> <ul><li>når du ble syk</li><li>når legen din regner med at du vil bli frisk</li></ul>',
+
+  'dokumentasjon.etablererEgenVirksomhet.tittel':
+    'Næringsfaglig vurdering av virksomheten du etablerer',
+  'dokumentasjon.etablererEgenVirksomhet.beskrivelse':
+    'Får du ikke dagpenger mens du etablerer egen virksomhet, må du skaffe næringsfaglig vurdering fra kommunen eller fylkeskommunen. Du kan også bruke en annen faglig kompetanse. ',
+
+  'dokumentasjon.tidligereFakturaer.tittel': 'Tidligere fakturaer',
+  'dokumentasjon.tidligereFakturaer.beskrivelse':
+    'Hvis du søker stønader tilbake i tid, må du legge ved tidligere fakturaer.',
+
+  'dokumentasjon.barnehageOgLiknende.tittel':
+    'Faktura fra barnepassordningen for perioden du søker om nå',
+  'dokumentasjon.barnehageOgLiknende.beskrivelse':
+    'Vi må ha ny faktura for hvert barnehage-/skoleår. Det er ikke tilstrekkelig at du har levert faktura for tidligere år.' +
+    '<br />' +
+    'Fakturaen må stå i ditt navn, og utgifter til opphold må være spesifisert.' +
+    '<br />' +
+    'Husk å legge ved tidligere fakturaer hvis du søker stønad tilbake i tid.',
+  'dokumentasjon.privatBarnepass.tittel': 'Avtalen du har med barnepasseren',
+  'dokumentasjon.privatBarnepass.beskrivelse':
+    'Avtalen må inneholde:\n<ul>' +
+    '<li>ditt navn, fødseslnummer og adresse</li>' +
+    '<li>barnepasserens navn og adresse</li>' +
+    '<li>navn og fødselsnummer for barnet/barna som passes</li>' +
+    '<li>perioden avtalen gjelder for</li>' +
+    '<li>dager i uken og klokkeslett for barnepass</li>' +
+    '<li>hvor mye du betaler for barnepass per måned, ikke inkludert kostpenger</li>' +
+    '<li>signaturer fra deg og barnepasser</li></ul>',
+  'dokumentasjon.trengerMerPassEnnJevnaldrede.tittel':
+    'Dokumentasjon på at barnet ditt har behov for vesentlig mer pass enn det som er vanlig for jevnaldrende',
+  'dokumentasjon.trengerMerPassEnnJevnaldrede.beskrivelse':
+    'Det dokumenterer du med uttalelse fra lege, spesialist eller annet helsepersonell.',
+  'dokumentasjon.barnepassArbeidstid.tittel':
+    'Dokumentasjon på at du må være borte fra hjemmet i lengre perioder på grunn av jobb',
+  'dokumentasjon.barnepassRoterendeArbeidstid.tittel':
+    'Dokumentasjon på at du jobber turnus eller skift, og jobber på tider utenom vanlig arbeidstid',
+  'dokumentasjon.ettersend.tittel':
+    'Fikk du ikke lastet opp all dokumentasjonen?',
+  'dokumentasjon.ettersend.tekst.barnetilsyn':
+    'Du får raskere svar på søknaden din hvis vi har all dokumentasjonen når vi starter behandlingen. Hvis du ikke fikk lastet opp alt nå, kan du <a href="https://www.nav.no/soknader/nb/person/familie/enslig-mor-eller-far/NAV%2015-00.02/ettersendelse/dokumentinnsending">ettersende det som mangler</a>.',
+  'dokumentasjon.ettersend.tekst.overgangsstønad':
+    'Du får raskere svar på søknaden din hvis vi har all dokumentasjonen når vi starter behandlingen. Hvis du ikke fikk lastet opp alt nå, kan du <a href="https://www.nav.no/soknader/nb/person/familie/enslig-mor-eller-far/NAV%2015-00.01/ettersendelse/dokumentinnsending">ettersende det som mangler</a>.',
+  'dokumentasjon.ettersend.tekst.skolepenger':
+    'Du får raskere svar på søknaden din hvis vi har all dokumentasjonen når vi starter behandlingen. Hvis du ikke fikk lastet opp alt nå, kan du <a href="https://www.nav.no/soknader/nb/person/familie/enslig-mor-eller-far/NAV%2015-00.04/ettersendelse/dokumentinnsending">ettersende det som mangler</a>.',
+
+  'kvittering.takk': 'Takk for søknaden',
+  'kvittering.alert.mottatt': 'Søknaden din om overgangsstønad er mottatt',
+  'kvittering.skolepenger.alert.mottatt':
+    'Søknaden din om stønad til skolepenger er mottatt',
+  'kvittering.barnetilsyn.alert.mottatt':
+    'Søknaden din om stønad til barnetilsyn er mottatt',
+  'kvittering.tekst.arbeidssøker':
+    'Husk å registrere deg som arbeidssøker hos NAV',
+  'kvittering.knapp.arbeidssøker': 'Registrer deg som arbeidssøker',
+  'kvittering.tekst.altViTrenger':
+    'Spørsmålene i søknaden, sammen med eventuell dokumentasjon, gir oss svar på det vi trenger for å behandle søknaden din. Hvis vi trenger mer informasjon vil vi ta kontakt med deg.',
+  'kvittering.tekst.dineSaker':
+    'Du kan finne søknaden din i <a href="https://www.nav.no/no/ditt-nav">Dine saker</a> når vi starter å behandle den. Finn <a href="https://www.nav.no/no/nav-og-samfunn/om-nav/saksbehandlingstider-i-nav">saksbehandlingstiden</a> for ditt fylke.',
+  'kvittering.knapp.dineSaker': 'Skriv ut kvittering',
+  'kvittering.tittel.skolepenger':
+    'Som enslig mor eller far under utdanning, kan du ha rett til stønad til skolepenger',
+  'kvittering.tekst.skolepenger':
+    'Hvis vi kommer frem til at denne utdanningen er nødvendig for at du skal kunne komme i jobb og forsørge deg selv, kan du ha rett til skolepenger. Det vil si studieavgift, semesteravgift og eksamensgebyr.',
+  'kvittering.lenke.skolepenger': 'Les mer om stønad til skolepenger',
+  'kvittering.knapp.skolepenger': 'Søk stønad til skolepenger',
+
+  'kvittering.tittel.tarUtdanning':
+    'Som enslig mor eller far under utdanning, kan du ha rett til stønad til skolepenger',
+  'kvittering.beskrivelse.tarUtdanning':
+    'Hvis vi kommer frem til at denne utdanningen er nødvendig for at du skal kunne komme i jobb og forsørge deg selv, kan du ha rett til skolepenger. Det vil si studieavgift, semesteravgift og eksamensgebyr.',
+  'kvittering.lenke.tarUtdanning': 'Les mer om stønad til skolepenger',
+  'kvittering.knapp.tarUtdanning': 'Søk stønad til skolepenger',
+  'kvittering.tittel.tilleggsstønader':
+    'Du kan også ha rett til tilleggsstønader når du er under utdanning',
+  'kvittering.beskrivelse.tilleggsstønader':
+    'Stønadene kan dekke utgifter til <ul>' +
+    '<li>barnepass</li>' +
+    '<li>læremidler</li>' +
+    '<li>daglig reise</li>' +
+    '<li>reise til samling</li>' +
+    '<li>reise på grunn av oppstart, avslutning eller hjemreise</li>' +
+    '<li>bolig og overnatting</li>' +
+    '<li>flytting</li> </ul>' +
+    '<br/> <i>NB! Du kan kun få tilleggsstønader hvis du kvalifiserer til overgangsstønad</i>',
+  'kvittering.lenke.tilleggsstønader': 'Les mer om tilleggsstønader',
+  'kvittering.knapp.tilleggsstønader': 'Søk tilleggsstønader',
+
+  'kvittering.tittel.tilleggsstønader.arbeidssøker':
+    'Som enslig mor eller far som søker arbeid, kan du ha rett til tilleggsstønader',
+  'kvittering.beskrivelse.tilleggsstønader.arbeidssøker':
+    'Stønadene kan dekke utgifter til <ul>' +
+    '<li>barnepass</li>' +
+    '<li>flytting</li></ul>' +
+    '<i>NB! Du kan kun få tilleggsstønader hvis du kvalifiserer til overgangsstønad</i>',
+  'kvittering.lenke.tilleggsstønader.arbeidssøker':
+    'Les mer om tilleggsstønader',
+  'kvittering.knapp.tilleggsstønader.arbeidssøker': 'Søk tilleggsstønader',
+
+  'kvittering.tittel.tilleggsstønader.aktivitetskrav':
+    'Som enslig mor eller far i arbeid, kan du ha rett til stønad til barnetilsyn',
+  'kvittering.beskrivelse.tilleggsstønader.aktivitetskrav':
+    'Stønaden dekker 64 prosent av utgiftene til barnepass, som for eksempel barnehage, skolefritidsordning (SFO) eller dagmamma. <br/><br/> ' +
+    'Dersom du tjener mer enn 608 106 kroner i året, har du ikke rett til stønaden.<br/><br/>' +
+    'Du kan som hovedregel motta stønaden frem til barnet fullfører fjerde skoleår.',
+  'kvittering.lenke.tilleggsstønader.aktivitetskrav':
+    'Les mer om stønad til barnetilsyn',
+  'kvittering.knapp.tilleggsstønader.aktivitetskrav':
+    'Søk stønad til barnetilsyn',
+  'overgangsstønad.feilsituasjon':
+    'Vi har dessverre tekniske problemer, så du får ikke sendt inn søknaden akkurat nå. Vennligst prøv igjen litt senere. Vi beklager ulempen.',
+};
