+++ src/utils/språk.ts
@@ -1,34 +1,5 @@
-import tekster from '../language/tekster';
-import { Språk } from '../models/felles/språk';
-
-export const getMessages = (locale: string) => {
-  if (locale === 'en') {
-    return tekster.en;
-  } else if (locale === 'nn') {
-    return tekster.nn;
-  } else {
-    return tekster.nb;
-  }
-};
-
 export const førsteBokstavStor = (tekst: string) => {
   return tekst.charAt(0).toUpperCase() + tekst.slice(1);
-};
-
-export const hentListeMedSpråk = (): Språk[] => [
-  { tittel: 'Bokmål', locale: 'nb' },
-  { tittel: 'Nynorsk', locale: 'nn' },
-  { tittel: 'English', locale: 'en' },
-];
-
-export const hentValgtSpråk = (locale: string) => {
-  let språk: string = '';
-  locale === 'en'
-    ? (språk = 'English')
-    : locale === 'nn'
-    ? (språk = 'Nynorsk')
-    : (språk = 'Bokmål');
-  return språk;
 };
 
 export const hentBeskjedMedNavn = (navn: string, tekststreng: string) => {
